source/requirements.txt
pikepdf==9.0.0
Pillow==10.3.0
